# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HeyCreator API",
    "description": "Influencer discovery and campaign management platform API. Connects brands with creators across Instagram, TikTok, YouTube, Twitter, and Facebook.",
    "version": "1.0.0",
    "contact": {
      "name": "HeyCreator Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/auth/resend-verification": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Resend email verification",
        "description": "Resends the email verification link to the user. Requires the user's Firebase UID.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["uid"],
                "properties": {
                  "uid": { "type": "string", "description": "Firebase UID of the user" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/auth/instagram/callback": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Instagram OAuth callback",
        "description": "Handles Instagram OAuth callback. Exchanges authorization code for access token, creates or updates user in Firebase, and returns a custom token.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": { "type": "string", "description": "Instagram authorization code from OAuth redirect" },
                  "role": { "$ref": "#/components/schemas/UserRole" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "customToken": { "type": "string", "description": "Firebase custom token for client-side sign-in" },
                    "instagramData": { "$ref": "#/components/schemas/InstagramBusinessAccount" },
                    "isNewUser": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID Token",
        "description": "Firebase Authentication ID token. Obtain by calling `firebase.auth().currentUser.getIdToken()`."
      },
      "UserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Development fallback: Firebase UID passed directly as header. Not for production use."
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": ["influencer", "brand"]
      },
...
      "InstagramBusinessAccount": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "name": { "type": "string" },
          "profile_picture_url": { "type": "string" },
          "followers_count": { "type": "integer" },
          "follows_count": { "type": "integer" },
          "media_count": { "type": "integer" },
          "biography": { "type": "string" },
          "website": { "type": "string" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "type": "object" },
          "count": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — validation error or missing fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
}